+++ frontend/package.json
@@ -21,6 +21,7 @@
   },
   "devDependencies": {
     "@ant-design/icons": "^4.2.2",
+    "@types/react-router-dom": "^5.1.6",
     "@typescript-eslint/eslint-plugin": "^4.6.0",
     "@typescript-eslint/parser": "^4.6.0",
     "antd": "^4.8.0",
@@ -28,7 +29,8 @@
     "eslint-plugin-prettier": "^3.1.4",
     "husky": "^4.3.0",
     "lint-staged": "^10.5.0",
-    "prettier": "^2.1.2"
+    "prettier": "^2.1.2",
+    "react-router-dom": "^5.2.0"
   },
   "browserslist": {
     "production": [
