+++ frontend/src/components/navigation/SearchBar.tsx
@@ -5,14 +5,14 @@ const { Search } = Input;
 
 const styles: { [identifier: string]: CSSProperties } = {
   search: {
-    width: '40%'
-  }
-}
+    width: "40%",
+  },
+};
 
 export default function SearchBar(): ReactElement {
-  const onSearch = (value: string) => console.log(value);
-  
+  const onSearch = (value: string) => alert(value);
+
   return (
     <Search placeholder="Search..." onSearch={onSearch} style={styles.search} />
-  )
+  );
 }
